% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplsAux.R
\name{coef.cv.gpls}
\alias{coef.cv.gpls}
\title{Extract coefficients from a cv.gpls model}
\usage{
\method{coef}{cv.gpls}(object, which = "lambda.1se", ...)
}
\arguments{
\item{object}{A cv.gpls object}

\item{which}{a character string, for which lambda shoudl coefficients be returned}

\item{...}{further arguments, currently ignored}
}
\value{
The vector of coefficients
}
\description{
Extract coefficients from a cv.gpls model
}
