% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penglsAux.R
\name{coef.cv.pengls}
\alias{coef.cv.pengls}
\title{Extract coefficients from a cv.pengls model}
\usage{
\method{coef}{cv.pengls}(object, which = "lambda.1se", ...)
}
\arguments{
\item{object}{A cv.pengls object}

\item{which}{a character string, for which lambda shoudl coefficients be returned}

\item{...}{further arguments, currently ignored}
}
\value{
The vector of coefficients
}
\description{
Extract coefficients from a cv.pengls model
}
