% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplsAux.R
\name{coef.gpls}
\alias{coef.gpls}
\title{Extract coefficients from a gpls model}
\usage{
\method{coef}{gpls}(object, ...)
}
\arguments{
\item{object}{A gpls object}

\item{...}{further arguments, currently ignored}
}
\value{
The vector of coefficients
}
\description{
Extract coefficients from a gpls model
}
