% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplsAux.R
\name{coef.pengls}
\alias{coef.pengls}
\title{Extract coefficients from a pengls model}
\usage{
\method{coef}{pengls}(object, ...)
}
\arguments{
\item{object}{A pengls object}

\item{...}{further arguments, currently ignored}
}
\value{
The vector of coefficients
}
\description{
Extract coefficients from a pengls model
}
