% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLoss.R
\name{getLoss}
\alias{getLoss}
\title{Calculate the loss given predicted and observed values}
\usage{
getLoss(preds, obs, loss)
}
\arguments{
\item{preds}{Matrix of predicted values}

\item{obs}{vector of observed values}

\item{loss}{a character vector indicating the loss type, see ?cv.pengls}
}
\value{
the evaluated loss
}
\description{
Calculate the loss given predicted and observed values
}
