% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpls.coef.R
\name{gpls.coef}
\alias{gpls.coef}
\title{Extract coefficients from a gpls model}
\usage{
\method{coef}{gpls}(object, ...)
}
\arguments{
\item{object}{A gpls object}

\item{...}{}
}
\description{
Extract coefficients from a gpls model
}
