% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpls.predict.R
\name{gpls.predict}
\alias{gpls.predict}
\title{Make predictions from a gpls model}
\usage{
\method{predict}{gpls}(object, ...)
}
\arguments{
\item{object}{A gpls object}
}
\description{
Make predictions from a gpls model
}
