% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penglsAux.R
\name{predict.cv.pengls}
\alias{predict.cv.pengls}
\title{Make predictions from a cv.pengls model}
\usage{
\method{predict}{cv.pengls}(object, ...)
}
\arguments{
\item{object}{A cv.pengls object}

\item{...}{further arguments, currently ignored}
}
\value{
A vector with predicted values
}
\description{
Make predictions from a cv.pengls model
}
