% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplsAux.R
\name{predict.gpls}
\alias{predict.gpls}
\title{Make predictions from a gpls model}
\usage{
\method{predict}{gpls}(object, ...)
}
\arguments{
\item{object}{A gpls object}

\item{...}{further arguments, currently ignored}
}
\value{
A vector with predicted values
}
\description{
Make predictions from a gpls model
}
