% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penglsAux.R
\name{predict.pengls}
\alias{predict.pengls}
\title{Make predictions from a pengls model}
\usage{
\method{predict}{pengls}(object, ...)
}
\arguments{
\item{object}{A pengls object}

\item{...}{further arguments, currently ignored}
}
\value{
A vector with predicted values
}
\description{
Make predictions from a pengls model
}
