% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplsAux.R
\name{print.cv.gpls}
\alias{print.cv.gpls}
\title{Print a summary of a cv.gpls model}
\usage{
\method{print}{cv.gpls}(x, ...)
}
\arguments{
\item{x}{A cv.gpls object}

\item{...}{further arguments, currently ignored}
}
\value{
Prints output to console
}
\description{
Print a summary of a cv.gpls model
}
