% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplsAux.R
\name{print.cv.pengls}
\alias{print.cv.pengls}
\title{Print a summary of a cv.pengls model}
\usage{
\method{print}{cv.pengls}(x, ...)
}
\arguments{
\item{x}{A cv.pengls object}

\item{...}{further arguments, currently ignored}
}
\value{
Prints output to console
}
\description{
Print a summary of a cv.pengls model
}
