% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplsAux.R
\name{print.gpls}
\alias{print.gpls}
\title{Print a summary of a gpls model}
\usage{
\method{print}{gpls}(x, ...)
}
\arguments{
\item{x}{A gpls object}

\item{...}{further arguments, currently ignored}
}
\value{
Prints output to console
}
\description{
Print a summary of a gpls model
}
