% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplsAux.R
\name{print.pengls}
\alias{print.pengls}
\title{Print a summary of a pengls model}
\usage{
\method{print}{pengls}(x, ...)
}
\arguments{
\item{x}{A pengls object}

\item{...}{further arguments, currently ignored}
}
\value{
Prints output to console
}
\description{
Print a summary of a pengls model
}
